Upgrade ansible-docker to override keyserver
Sometimes, the default docker keyserver is not available and
causes provisioning to fail. When this happens, we can point
to a different keyserver and continue provisioning. This new
version of the ansible-docker role allows overriding default
keyserver by adding a line like this to group_vars/all:

    docker_keyserver: "hkp://pgp.mit.edu:80"

See https://github.com/azavea/ansible-docker/pull/8