Keep modification integrity as drawn
In MMW, modifications were not burned into the raster before spatial
operations, so the intersections and their order were calculated and
clipped on the client, resulting in modification "pieces".  In the
pollinator implementation, overlaps will be burned in using the order
they were applied.  Additionally, since the model actually takes a
buffer around the AoI, we specifically allow modification outside of
the delineated AoI.