Add script to aggregate crop values
Converts the CSV of raw CDL values to an aggreggated list of
just the unique values of `Attributes`, assigning new ids to
represent the groups and retaining the static HN and HF mean
values per group.

Additionally creates a json mapping file that can be used to
reclassify CDL codes from the CDL raster to the new grouped
codes.