Convert yield to yield per acre
Since comparisons are made against yield types across scenaios where
the specific amount of crop planted may change, we are converting the
yield as calculated for a parituclar field to that of yield per acre of
crop type on a particular field.  That way if the planted crop area is
reduced, the yield per acre should be reduced in tandem, even if the
chagne wasn't due to pollinator abundance.