Ensure results are saved when set
The project was previously saved after polling ended, leading to
situations where a saved project, reloaded, would have to poll again
for results since the inputmod_hash on the saved results did not match
the hash saved on the scenario. Now it will save when results are set,
no matter the source.