Use model changes to update results view
Previously, the ResultsView decided to show the error, polling, or results based
on the outcome of a promise that is created in the initialize method. This means
that once the promise has failed, the results can never be displayed in a
subsequent run of the model. This problem manifests in the following situation:
when polling for initial results, if the user makes a modification to the
scenario, the initial run is cancelled which causes the error message to
display, and the new results never show up. This bug is an instance of a more
general pattern I've noticed in the app which is using promises to trigger
renders in views. In general, we should strive to use a single mechanism to
trigger UI updates, namely Backbone model events.

To fix this problem, I made the taskHelper in ScenarioModel set a new model
attribute ("pollError") when there is a polling failure. This is so we can pass
the error through the ScenarioModel to the view that renders the error. I also
pushed the code which decides whether to render the error, polling, or results
down the component hierarchy into the ResultsTabContentView, since that decision
needs to be made individually, for each scenario. The decision is now based on
the state of the model, so it's possible to go back and forth between polling,
error and results.