Add additional colormap entries
Since the total list of crops has expanded, so too must the color map
gain entries for the new additions. Since the work to expand the list
has resulted in shifting the group ids, the colors will need to get
reworked. For that reason, this commit just repeats existing colors.