Add additional crop types
These crops do not appear in the CDL raster, so the IDs will never
appear in it. However, they will be available as options for
"correcting" the CDL raster at run time, so they need to have values
for the model to operate on. These entries were provided by ICP.