Use data edits to CDL instead of scripted changes
Previously, snythetic crop types were added programatically to the base
CDL. Instead, these are now appended directly to the base CDL csv and
we no longer need to update values in this script. The previous commit
was not reverted since we still want the change that allow writing out
the new density value to the grouped csv.