Use same project for Compare view as Modeling
Previously, we would create a sandboxed copy of the main
project for the Compare view. This was done in order to
preserve the state of the main project from any changes
made in the Compare view, which could trigger model runs
and modify input values.

This led to bugs such as spinners continuing to spin if
that's how Compare view was initialized, even after any
polling finishes. The original project's results would
be updated, but not of the sandboxed copy, causing the
user to see stale results.

Since we don't expect the Pollinator App to be able to
trigger model runs from the Compare view, it is safe to
remove the sandboxed copy of the project and just use the
original one. This simplifies the code, making it easier
to reason about, as well as removing any bugs like the
one mentioned above.

This functionality was originally added in Model My Watershed
https://github.com/WikiWatershed/model-my-watershed/pull/673