Refactor fetching logic while keeping behavior
Previously the checks for when to fetch were tangled with
other pieces, making it difficult to trigger project-level
fetches after the first time.

By moving the event listeners to `ProjectModel.initialize`,
we ensure that they are attached only once, at initialization,
and not in every call of `ProjectModel.fetchResultsIfNeeded`.
This further reduces the responsibility of `ProjectModel.
fetchResultsIfNeeded` to a single task, thus increasing its
reusability.

We remove the check for `active` from within `ScenarioModel.
fetchResultsIfNeeded` to without. This makes it possible to
fetch results of scenarios when none are active, such as from
the Compare view. This check is moved outside, so that only
the active scenario's results will be fetched when loading
the Model view.

This should not change the actual user interaction at all.