Make crop layer button more clickable
Previously, one had to click the tiny eye button to enable or
disable the crop layer. Now one can click the label to the same
effect.

While the issue asked for the entire button to be clickable,
only the label has been made clickable because once the layer is
enabled there are a number of other controls in the button, such
as the opacity slider and the legend toggle, which require click
behaviors of their own, which would conflict with that of the
button if it had one has a whole.

Refs #183 :lipstick: