Add Docker
To allow us to have a more modern toolchain for the new Beekeepers
app, we need to have a more modern version of Node. Upgrading the
existing version could be a very large undertaking though, so we
will use a newer version within a Docker container. For that, we
first need to install Docker.