Configure Webpack to work with Django
Add webpack-bundle-tracker plugin that generates a webpack-stats
file, which is read by Django to reload JavaScript at runtime.

Also allow CORS, since the JavaScript will be loaded via webpack-
dev-server, which is running on a different port than Django.

Make it so this only happens in development, and there is no hot
reloading in production.