Add beekeepers script to run yarn
This beekeepers script runs webpack via Docker within the VM. It
is a simple invocation of the yarn.sh script for beekeepers.

For generating the production bundle, use

    ./scripts/beekeepers.sh run bundle

For running the dev server, use

    ./scripts/beekeepers.sh run dev-server

The dev server runs on :32579, but the developer can just go to
:8000/?beekeepers as usual and the JavaScript will be picked up
intelligently.

After the dev-server is stopped, the production bundle must be
regenerated.