Add provisioner for Beekeepers
The provisioner does the following:

  - Allows vagrant to run docker by adding the vagrant user to the
    docker group
  - Fetches node docker image
  - Installs Beekeepers NPM dependencies
  - Generates the Beekeepers bundle files
  - Creates a symlink of the bundle output directory to the static
    directory to allow access from the web

It would be nice to reuse the yarn.sh script instead of specifying
docker commands inline. However, the yarn.sh script must be able
to take input on a terminal for stopping webpack-dev-server, so we
specify -it in yarn.sh. Unfortunately, since ansible is not a TTY,
we cannot use the script through it.