Downgrade psycopg2 due to OpenSSL inconsistency
The wheel package for `psycopg2` is not compatible with the Python `ssl`
module due to their dependency on different versions of OpenSSL (the
wheel packages its own version, while `ssl` uses the system version).

Attempts to resolve by installing the dependency from source proved
challenging because we install `psycopg2` via Ansible
(`postgresql-support`) in a way that makes it difficult to supply the
`--no-binary` override for `pip`. Instead, we've downgraded to a version
prior to wheel support.

Lastly, the `libpq5` and `libpq-dev` dependencies needed to be
downgraded as well. This is because `psycopg2==2.6.2` is unable to make
sense of a `libpq` version greater than 9.

See: https://github.com/psycopg/psycopg2/issues/543