Mount AWS credentials
AWS credentials for bees is needed to make requests to beekeepers rasters housed in s3. In order for the credentials to be available to the user which is the django app in the VM, needs to have chmod permissions 644 and for owner to be set manually to the app user, 'icp'.

Refs #288