Mount AWS details to the env for Django with Ansible
Set the aws info to use from the credentials and s3 bucket such that that information is avail globally to the django app. Boto3 under the hood of rasterio picked up the AWS_PROFILE env variable automatically and authorizes a connection to AWS. This setup will need further configuration to accomodate staging and production credentials.

Refs #288