Add Sidebar component
The new Sidebar component connects to the redux state, and if
the number of apiaries is 0, shows the Splash page. Otherwise,
it shows the Prototype, which has been adjusted to be just the
contents of the sidebar.

The Map component is currently a placeholder, but has been
extracted from the Prototype. App has been updated accordingly.