Separate saved reducer for persistence
The `main` reducer will have a number of transient values that
do not need to be saved, and its shape will be changing in the
early days of development quite frequently, making saving state
inconvenient. It is imprudent to persist its contents at this
time.

One alternative is to have a separate top-level reducer to
explicitly hold items that should persist. For this, the place-
holder `saved` reducer is added. It is currently empty, and
will be populated with actions and state in the future.

This allows us to have redux-persist setup and ready to go
without impeding the evolution of main state.