Add login and sign up modals with working actions
Reactjs-popup library's modals are such that templates cannot be slotted into a single modal if you want to hook into the modal "close" action, as we do. The most sensible way was to rely on redux state but we will not want to persist the modal state values in localStorage.

Refs #325