Move crop types and colors to Beekeepers app
Both Beekeepers and Pollination Mappper need access to the same
set of values and colors for the Crop Data Layer legend. We can
duplicate the values, but that would lead to redundancy and the
need to explicitly update both sets of files.

Instead, we can refer to the same file from two places. Ideally
the Beekeepers app would refer to the files in their original
location. Unfortunately, since it is built using Docker which
is only aware of the Beekeepers directory, we cannot reference
files outside this directory.

Thus, the two files are moved to be within the Beekeepers
directory, and the original locations are updated to be
symlinked. Since the original is built in Vagrant which has
access to the entire folder, this does not affect that build
process at all.