Show Log Out button on hover, log out on click
Adds a downward triangle ▾ icon next to the username shown when
a user is logged in. When the user hovers on their  username,
a hidden dropdown menu is shown, which currently has only one
button, "Log Out", clicking which logs the user out.

When the user is logged out, the apiaries are reset and the app
subsequently reverts to the initial state.