Save client-side apiaries for new users
If a user has created apiaries without logging in, and then
they do log in, we see if they have previously saved apiaries.
If they do, then the unsaved work is discarded. If they do not,
then the unsaved work is saved to the server.

This is useful for users who are trying out the app, and at
some point want to create an account and save their work.

Refs #328