Add helper functions for getting surveys
These surveys represent both actual and potential surveys,
with an additional "completed" key set to true when the survey
has been completed, and to false otherwise.

Also add one that will sort a list of survey in descending order.
Useful, since they are shown in descending order in the UI.

Also add another month to name translator, this time an object
that takes the two digit representation and converts to name.

Also adds an internal method for returning a matching survey in
a given list if found by month_year, or undefined.