Move survey requests to component
A survey only wants to be seen by its respective survey form modal, and not the rest of the app. While an ajax call from a component is not a typical React-Redux pattern, the survey data will never need to be accessed by other components than when and where requested. Also there are no side effects as would otherwise be introduced by data persist in redux.

Refs #379