Update Survey Endpoint serialization
In the previous design, the subsurvey fields were top-level and
the survey fields were encapsulated within a `survey` key. This
makes validating different kinds of subsurveys hard, especially
for MONTHLY surveys which can have more than one subsurvey for
each survey.

In the new design, the survey fields are the top-level ones, with
the subsurvey fields encapsulated within an `april`, `november`,
or `monthlies` key. `monthlies` is an array, but the others are
singleton objects.

When creating new surveys, only the fields required for a given
kind of survey are needed, e.g. when the `survey_type` is set to
"APRIL", only the `april` key needs to be specified. The contents
of the subsurvey fields will be validated against their models
just as before.