Add Monthly Survey and Subsurvey forms
The main state is kept in the parent component, and we have
simple child components for each of the three tabs. When the
submit button is pressed, we compose a payload for every
colony which has a "name" specified, and send that to the
back-end.

Currently, after submission, the form does not update / populate
with disabled values as other forms do. This will most likely
be replaced with the success message from #416 soon.