Pin to pip 18.1
When pip 19.0 was released, builds started failing with messages similar to Cannot uninstall 'requests'. It is a distutils installed project and thus we cannot accurately determine which files belong to it which would lead to only a partial uninstall. When we stopped installing the python-requests-whl package, this error would move to another package, like urllib3. In the end, it seemed that pinning the pip version to 18.1 prevented these errors from happening.