Sort a deep copy of apiary array
Previously we would sort the original array, thus mutating
the redux state out of band, which was expected to be sorted
by the default arrangement by the code which calculates the
next apiary's marker. As the original array would be sorted
otherwise, the next marker calculation would falter.

By ensuring the sort happens strictly in the render code,
we guarantee the original array to remain untouched.