Only deep copy on non-default sort
Since the original array is in the default sort order already,
when the user wants it sorted by the default order we just use
the original array directly. In other cases, we make a deep
copy and sort that, so as not to alter the redux state out of
band.