Reset User Survey Form state on submit
Previously, we did not clear out the state of the User Survey
form when it was submitted. This led to previous values being
reused if a user were to sign in to one account, fill out the
survey, sign out, and sign in to another account, and fill it
out again.

To prevent this from happening, we now clear the local state
when the form is submitted.