Add ability to remove unfetched apiaries
Currently, if there is a server error while fetching apiary scores,
they remain on screen and cannot be removed. By adding the same
trash button to them that filled-in apiaries have, we allow the
user two exits from the state: to retry and to remove the apiary.

Also, the apiary's marker is also added to the card, so the user
can associate the malfunctioning API to the map.