Remove Docker deinstall from beekeepers provisioning
While we would like to be able to remove Docker from the app AMI
entirely as described in e104379a91abdb438bad70662aeefde936314d0e,
the Docker Ansible role raises an error at the end of the AMI build
if Docker is not installed. Undo the Docker deinstall step to get
AMI builds passing again.