Don't try to fetch results until project is loaded
In 1ddff11, as part of #165, the result fetching was refactored
to be a little more eager. This works well for most cases, but
when the results are artificially cleared, it would make existing
projects crash, because the results would try to fetch before the
project would finish loading (triggered by the project's fetch
method), causing the front-end to request the yield without
specifying an area of interest.

By adding a test to not fetch results until a project has loaded,
we ensure that the eager call for project's with reset results
do not fire. The results are fetched again when the result view
is loaded, by which time the project has loaded correctly, so
the project can reinitialize correctly.