Show survey results in the order they were saved
Even though the front-end sends the apiaries to the back-end
in the correct order, and they are saved in the correct order,
for some reason when they are queried for again, the serializer
sends them back in reverse order.

It was unclear how this can be remedied on the DRF side, as the
documentation did not provide any hints, nor could this question
be found on Stack Overflow.

In the interest of expediency, I've decided to sort the results
in the front-end when the Monthly Survey details modal is loaded.
This ensures they are shown in the correct order, regardless of
what order they are sent down from the API in.