Fix overlapping items in IE 11
In a number of places, we were using `display: flex` with
`flex-direction: column` for simple layout, which is in most
cases a no-op, since content flows in a columnar direction
out of the box.

By removing those declarations and compensating with some
strategic margins, we ensure the page looks good and works
across all browsers.