Use smart background color to mask layout issues
Due to the use of flexbox, there are cases where the Survey
page doesn't cover the entire page, only the content shown
in the viewport. The lighter background color was applied
only to this viewport area, so when the user would scroll in
IE or Firefox (and in some few cases, Chrome), they would see
the darker background color show from underneath.

Figuring out a cross-browser solution to these layout issues
could be challenging, so in the interest of expediency we
set the background to be the same lighter color as that of
the Survey page. This ensures that the break is never seen,
and this doesn't affect any other part of the app because
there is always content placed over the body.

Connects #465