Ensure API listing of apiaries is always ordered
Previously, for some unknown reason, the responses had
begun to be sorted in reverse. The front-end relies on
getting a response sorted in ascending order of the id,
and began to malfunction. By ensuring that the results
are always sorted in the correct order, we record this
assumption explicitly and ensure the front-end works.