Add Apiary table to data export
Previously, we only exported the Survey tables. This was not
enough, as the Apiary surveys could not be connected to the
User surveys, since the user_id was never attached to the
Apiary surveys.

By exporting the Apiary table along with the results, we give
the entire set of collected data to the exporting user, which
will give them the flexibility to connect and slice it in
many different ways.

Furthermore, the user_id column is also added to the Survey
table export, which can help make the connection without an
extra step of connecting with an Apiary first.