Enable editing apiary names
When a user hovers on an apiary name in the card, they see
a faint border an a pencil icon indicating that it can be
edited. Clicking either the apiary name or the pencil icon
makes the name editable. When the edits are complete, a
user may hit "Enter" or the check button to save the
changes. If there are no changes, nothing is done. Otherwise,
the changes are sent to the server immediately and saved to
the database.