Make input more accessible
In the previous implementation, the input element would be
replaced as a whole based on differing conditions. This is
bad for accessibility, since the old element which had the
focus would be replaced with a new one that does not. This
is alleviated by making all attributes conditional, and
using them directly in a single element.