Ensure reverse geocoding still works
Previously when the card was a simple functional component, the
displayed name was always sourced from the props. Now the prop
is used for the initialization, with an internal state variable
taking it from there.

This is required for the editing, but gets in the way of updating
the name with a reverse geocoded address, which kicks in later.
To allow for this, we override the internal state with an updated
prop value whenever it is different from the last prop value.
This is only expected to happen in the case of reverse geocoding,
so we should be safe from other unintentional side-effects of
overriding a state variable with a prop value.