Pan & zoom map based on redux state
New behavior:
- Zoom to location when selecting a geocoded location.
- Turn the apiary card marker into an interactive map button, as requested by the client

The new source of truth for the map view will be the app redux state rather than the self-managed leaflet map state. This is because components external to the map need to make updates to the map view (that is, zoom & center). Unify on redux for clarity of map behavior.

Refs #531