Change button for editing apiary names
Users had trouble interpreting the image buttons for edit and save
within the card component, so we replaced the icons with text to clarify
their purpose. Refactored the CardButton component export into two
higher-order components, CardButtonWithText and CardButtonWithIcon, to
allow reusability of the CardButton code as well as differing sets of
required props for the different button types. The components are
co-located in the CardButton file as named exports.