Ensure ICP_SERVICES_IP is set within the virtual machine
In a few areas of the Ansible configuration, we pull certain variable
values out of the environment (e.g., the IP of the virtual machine with
the database on it, so that the application can communicate with it).
When we moved to using the local Ansible provider with Vagrant, we lost
the ability to access virtual machine host environment variables,
because Ansible is executing within the virtual machine.

These changes make use of the extra_vars function provided by the
Vagrant Ansible local provider, which is executed on the virtual machine
host, and has access to its environment.