Fix Python libc segfault tied to psycopg2 installation
Something about the wheel for psycopg2==2.7.x loads external operating
system libraries in a way that is incompatible with the versions of
those libraries provided on Trusty. To avoid any issues with bundled
wheel binaries, the most recent non psycopg2-binary package is being
installed.

Seems related to https://github.com/psycopg/psycopg2-wheels/issues/2.

No reference to psycopg2 is made in requirements.txt because the
application expects psycopg2 to already be there (via Ansible and the
postgresql-support role).