Edit colony questions in April Survey
Adds a line that displays the number of colonies (`num_colonies`) the
user centered for that apiary in the November survey.

Adds a max value to the remaining colonies (`num_colonies`) field which
limits it to number of colonies entered in the November survey.

Adds the question - “How many new colonies did you add to this apiary
over the winter?” with a default answer of 0. This required the
addition of a field to the AprilSurvey model, `num_new_colonies`.
Existing April Surveys will have this new field marked `null`.

In the database, `num_colonies` represents the total number of colonies.
Therefore the data stored to this field is equal to the remaining
colonies plus the new colonies.