Map correctly centers after zoom
Previously, we were not updating the map's center when zooming. This
resulted in the map panning to the previous center after trackpad
zooms, which are intended to re-center the map at the point where the
user's cursor is when beginning the zoom. We are now correctly updating
the map's center to equal the zoom location after a zoom action ends.